refactor(core): extract stream processing and runtime ports
Introduce bitfun-agent-stream and move stream processing logic, fixtures, and
stream regression tests out of bitfun-core while keeping a compatibility wrapper
for the existing core import path.

Move shared AI error categorization and detail helpers into bitfun-core-types,
then narrow BitFunError's concrete dependency coupling at the core boundary.

Introduce bitfun-runtime-ports as a DTO and trait-only boundary for agent
submissions, dynamic tools, config reads, and session transcript reads, with
initial core adapters for the coordinator, config service, and tool registry.

Keep product feature defaults unchanged while reducing the compile surface needed
for stream-focused tests.
