+++ src/crates/agent-stream/tests/common/fixture_loader.rs
