+++ src/crates/agent-stream/tests/common/mod.rs
