+++ src/crates/agent-stream/tests/common/sse_fixture_server.rs
