+++ src/crates/agent-stream/tests/fixtures/stream/anthropic/extended_thinking.sse
