+++ src/crates/agent-stream/tests/fixtures/stream/anthropic/inline_think_text.sse
