+++ src/crates/agent-stream/tests/fixtures/stream/anthropic/interleaved_parallel_tool_use.sse
