+++ src/crates/agent-stream/tests/fixtures/stream/anthropic/malformed_content_block_delta.sse
