+++ src/crates/agent-stream/tests/fixtures/stream/gemini/function_call_string_args.sse
