+++ src/crates/agent-stream/tests/fixtures/stream/openai/inline_think_text.sse
