+++ src/crates/agent-stream/tests/fixtures/stream/openai/tool_args_snapshot_stop_reason.sse
