+++ src/crates/agent-stream/tests/fixtures/stream/openai/tool_args_split_with_usage.sse
