+++ src/crates/agent-stream/tests/fixtures/stream/openai/tool_call_missing_type_field.sse
