+++ src/crates/agent-stream/tests/fixtures/stream/openai/tool_id_only_orphan_filtered.sse
