+++ src/crates/agent-stream/tests/fixtures/stream/responses/malformed_function_call_arguments.sse
