+++ src/crates/runtime-ports/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "bitfun-runtime-ports"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+description = "Thin runtime ports for BitFun core decomposition"
+
+[lib]
+name = "bitfun_runtime_ports"
+crate-type = ["rlib"]
+
+[dependencies]
+async-trait = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
