+++ src/crates/runtime-ports/src/lib.rs
@@ -1,0 +1,195 @@
+//! Thin runtime ports for boundaries that currently cross service and agentic
+//! concrete implementations.
+//!
+//! This crate intentionally contains only DTOs and traits. It must not depend
+//! on concrete managers, platform adapters, `bitfun-core`, or app crates.
+
+use serde::{Deserialize, Serialize};
+
+pub type PortResult<T> = Result<T, PortError>;
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum PortErrorKind {
+    NotAvailable,
+    NotFound,
+    InvalidRequest,
+    PermissionDenied,
+    Cancelled,
+    Timeout,
+    Backend,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct PortError {
+    pub kind: PortErrorKind,
+    pub message: String,
+}
+
+impl PortError {
+    pub fn new(kind: PortErrorKind, message: impl Into<String>) -> Self {
+        Self {
+            kind,
+            message: message.into(),
+        }
+    }
+}
+
+impl std::fmt::Display for PortError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{:?}: {}", self.kind, self.message)
+    }
+}
+
+impl std::error::Error for PortError {}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct AgentSessionCreateRequest {
+    pub session_name: String,
+    pub agent_type: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub workspace_path: Option<String>,
+    #[serde(default, skip_serializing_if = "serde_json::Map::is_empty")]
+    pub metadata: serde_json::Map<String, serde_json::Value>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct AgentSessionCreateResult {
+    pub session_id: String,
+    pub agent_type: String,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct AgentSubmissionRequest {
+    pub session_id: String,
+    pub message: String,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub attachments: Vec<AgentInputAttachment>,
+    #[serde(default, skip_serializing_if = "serde_json::Map::is_empty")]
+    pub metadata: serde_json::Map<String, serde_json::Value>,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct AgentInputAttachment {
+    pub kind: String,
+    pub id: String,
+    #[serde(default, skip_serializing_if = "serde_json::Map::is_empty")]
+    pub metadata: serde_json::Map<String, serde_json::Value>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct AgentSubmissionResult {
+    pub turn_id: String,
+    #[serde(default)]
+    pub accepted: bool,
+}
+
+#[async_trait::async_trait]
+pub trait AgentSubmissionPort: Send + Sync {
+    async fn create_session(
+        &self,
+        request: AgentSessionCreateRequest,
+    ) -> PortResult<AgentSessionCreateResult>;
+
+    async fn submit_message(
+        &self,
+        request: AgentSubmissionRequest,
+    ) -> PortResult<AgentSubmissionResult>;
+
+    async fn resolve_session_agent_type(&self, session_id: &str) -> PortResult<Option<String>>;
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct DynamicToolDescriptor {
+    pub name: String,
+    pub description: String,
+    pub input_schema: serde_json::Value,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub provider_id: Option<String>,
+}
+
+#[async_trait::async_trait]
+pub trait DynamicToolProvider: Send + Sync {
+    async fn list_dynamic_tools(&self) -> PortResult<Vec<DynamicToolDescriptor>>;
+}
+
+pub trait ToolDecorator<Tool>: Send + Sync {
+    fn decorate(&self, tool: Tool) -> Tool;
+}
+
+#[async_trait::async_trait]
+pub trait ConfigReadPort: Send + Sync {
+    async fn get_config_value(&self, key: &str) -> PortResult<Option<serde_json::Value>>;
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SessionTranscriptRequest {
+    pub session_id: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub from_turn_id: Option<String>,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SessionTranscript {
+    pub session_id: String,
+    #[serde(default)]
+    pub messages: Vec<TranscriptMessage>,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct TranscriptMessage {
+    pub role: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub turn_id: Option<String>,
+    #[serde(default)]
+    pub content: serde_json::Value,
+}
+
+#[async_trait::async_trait]
+pub trait SessionTranscriptReader: Send + Sync {
+    async fn read_session_transcript(
+        &self,
+        request: SessionTranscriptRequest,
+    ) -> PortResult<SessionTranscript>;
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn port_error_display_keeps_kind_and_message() {
+        let error = PortError::new(PortErrorKind::NotAvailable, "coordinator missing");
+
+        assert_eq!(
+            error.to_string(),
+            "NotAvailable: coordinator missing".to_string()
+        );
+    }
+
+    #[test]
+    fn agent_submission_request_serializes_with_stable_camel_case() {
+        let request = AgentSubmissionRequest {
+            session_id: "session_1".to_string(),
+            message: "hello".to_string(),
+            attachments: Vec::new(),
+            metadata: serde_json::Map::new(),
+        };
+
+        let json = serde_json::to_value(request).expect("serialize request");
+
+        assert_eq!(json["sessionId"], "session_1");
+        assert_eq!(json["message"], "hello");
+        assert!(json.get("attachments").is_none());
+    }
+}
